Move hook to client/sites/hooks
